requirements.txt
streamlit>=1.28.0
langchain-google-genai>=1.0.0
langchain>=0.1.0
python-dotenv>=1.0.0
requests>=2.31.0
Pillow>=10.0.0
google-generativeai>=0.3.0
serpapi>=0.1.5
pandas>=2.0.0
streamlit-option-menu>=0.3.6
pydantic>=2.0.0
pydantic-settings>=2.0.0 